+++ imports/lib/data/skillsDiagram.ts
