+++ imports/ui/components/Accordion/index.ts
@@ -1,0 +1,2 @@
+import Accordion from "./Accordion";
+export default Accordion;
