+++ imports/ui/components/AnimatedImagesCircle/index.ts
@@ -1,0 +1,2 @@
+import AnimatedImageCircle from "./AnimatedImageCircle";
+export default AnimatedImageCircle;
