+++ imports/ui/template/Header/index.ts
@@ -1,0 +1,2 @@
+import Header from './Header';
+export default Header;
